+++ changelog.md
@@ -1,7 +1,8 @@
 * mkchromecast (0.2.8) **unreleased**:
 
     - Preferences and volume windows always on top.
-    - The node streaming has been updated to work with v6.2.1.
+    - The node streaming has been updated to work with v6.2.1. This improves
+      stability for macOS users.
 
 * mkchromecast (0.2.7) **released**: 2016/06/16